api/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.3
docutils==0.16
filelock==3.0.12
idna==2.10
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.1
Pillow==8.1.0
pipenv==2020.11.15
Pygments==2.7.3
pypiwin32==223
pytz==2021.1
pywin32==300
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
